Clean up meeting grids further
Fix the horizontal rules that are supposed to display underneath the
meeting page titles.

I'm not completely sure what the deal is with the "clearfix" elements
in the meeting list, but the logic that inserted them was incorrect
for multiple upcoming meetings regardless. I changed the clearfix
classes to be attached to the parent containers instead, as is
suggested in the docs:

https://getbootstrap.com/docs/5.3/helpers/clearfix/

I'm not certain that clearfix is even needed in the grid anyway,
since it uses flexbox rather than floating elements.